step release 10
[Features]

* Updates RabbitMQ to version 3.10.25 @itsouvalas
* Updates Erlang to 25.3.2.2 @itsouvalas

This release is part of the step releases that will allow an operator
to do a manual in-place upgrade from v1.1.0 to v1.2.0 onwards. If you
do not intend to do a manual upgrade or you are deploying for the first
time, you can use the latest version currently available.